refactor(game-night): consolidate useGameKbStatus hooks into single source
Delete legacy hook apps/web/src/hooks/use-game-kb-status.ts and migrate
its 2 consumers (game-detail-mobile, GameTableZoneKnowledge) to the
normalized domain hook at lib/domain-hooks/useGameKbStatus.

Changes:
- Extend normalized hook with suggestedQuestions (previously only in
  legacy) to achieve feature parity.
- Remove the 'normalized' queryKey workaround now that there's no
  cache-key collision.
- Accept gameId: string | null | undefined for backward compatibility
  with legacy consumers that passed undefined.
- Update tests (5 → 5, added suggestedQuestions assertions + undefined
  gameId compatibility test).
