+++ apps/web/src/lib/domain-hooks/useGameKbStatus.ts
@@ -11,27 +11,29 @@ interface UseGameKbStatusResult {
   isIndexed: boolean;
   documentCount: number;
   coverageLevel: CoverageLevel;
+  suggestedQuestions: readonly string[];
   isLoading: boolean;
   error: Error | null;
 }
 
+const EMPTY_SUGGESTED_QUESTIONS: readonly string[] = Object.freeze([]);
+
 /**
  * Hook to check whether a game has an indexed knowledge base (PDF + RAG).
  *
- * Returns safe defaults (`isIndexed=false`, `documentCount=0`, `coverageLevel='None'`)
- * when `gameId` is null, when the backend returns null, or while fetching.
+ * Returns safe defaults (`isIndexed=false`, `documentCount=0`, `coverageLevel='None'`,
+ * empty `suggestedQuestions`) when `gameId` is null, when the backend returns null,
+ * or while fetching.
  *
- * Used by the Game Night wizard to surface a soft "KB available" hint
- * (MVP hardening F1 — PDF-aware filter).
+ * Used by:
+ *  - Game Night wizard soft "KB available" hint (F1)
+ *  - Library game detail page & game-table knowledge zone
  */
-export function useGameKbStatus(gameId: string | null): UseGameKbStatusResult {
+export function useGameKbStatus(gameId: string | null | undefined): UseGameKbStatusResult {
   const query = useQuery({
-    // 'normalized' suffix avoids cache-key collision with the legacy
-    // hook at @/hooks/use-game-kb-status which uses ['game-kb-status', gameId]
-    // and a different staleTime (5 minutes vs our 60s).
-    queryKey: ['game-kb-status', 'normalized', gameId],
+    queryKey: ['game-kb-status', gameId],
     queryFn: () => api.knowledgeBase.getUserGameKbStatus(gameId as string),
-    enabled: gameId !== null,
+    enabled: gameId != null,
     staleTime: 60_000,
   });
 
@@ -39,6 +41,7 @@ export function useGameKbStatus(gameId: string | null): UseGameKbStatusResult {
     isIndexed: query.data?.isIndexed ?? false,
     documentCount: query.data?.documentCount ?? 0,
     coverageLevel: query.data?.coverageLevel ?? 'None',
+    suggestedQuestions: query.data?.suggestedQuestions ?? EMPTY_SUGGESTED_QUESTIONS,
     isLoading: query.isLoading,
     error: query.error as Error | null,
   };
